vinyl: return maximal LSN from compact_new()
In the single mem concept, each range on the level zero has
some infiniruns and stores count of them. But when we start
compaction, we can't determine which infiniruns on level zero
intersects with the compacting range and we will add all of them
with max_lsn >= range->max_used_lsn. It means, that compact_priority
is not always equal to count of actually compacted ranges.
To know which infiniruns we have compacted in compact_complete(),
the max_compact_lsn can be used. Max_compact_lsn - the maximal LSN
of compacted infiniruns.

      +-------------------+
      |                   |  infinirun
      +-------------------+
+-------------------+
|                   |  infinirun
+-------------------+
              +-------------------+
              |                   |  infinirun
              +-------------------+
+-----------+
|   range   |
+-----------+

"On the fly" we can't fast determine, which runs need to add
to compaction, so we add all of them and inside the write iterator
in the worker not needed runs will be skipped.