vinyl: implement deferred restore of runs and mem
The infiniruns and mems in the single mem concept haven't
ranges and behaviour of their iterators must be like txw_iterator.
That is we must restore them in the read iterator when
the next range is got for iteration.

But this increase the memory lookups and disk reads for such
restorations, because the index->cache sometimes already
contains the needed statements and we can avoid restoration.

That is, we must allow to the cache iterator to try to find the
statement before restoration of mems and runs. So the run and
mem iterator must provide the ability for deferred restoration.

The idea is to remember the statement for restoration and
make restore only if the cache iterator returned nothing.