vinyl: use tuple instead of raw msgpack for range begin and end
Maintaining range->{begin,end} as tuples is useful for the concept of
run slices, which is introduced by the following patches. A slice may
inherit its begin and end from a range, so basically we have two
alternatives: either copy keys or take references. The latter seems to
be more straightforward.