vinyl: add slice size estimate
We use run->info.keys to estimate the size of a new run's bloom filter.
We use run->info.size to trigger range split/coalescing. If a range
contains a slice that spans only a part of a run, we can't use run->info
stats, so this patch introduces the following slice stats: number of
keys (for the bloom filter) and the size on disk (for split/coalesce).
These two counters are not accurate, they are only estimates, because
calculating exact numbers would require disk reads. Instead we simply
take the corresponding run's stat and multiply it by

    slice page count / run page count