vinyl: fix txw iterator for empty key
Fixes commit 976b31cb9 ("vinyl: fix order error in txv_iterator_start").

If search key is empty, txw iterator starts from the first or last entry
in the write set depending on the iterator direction. This is incorrect,
because the write set is grouped by index so if the first/last entry
happens to be for another index, txw iterator will stop immediately even
if there are statements for the given index in the write set. Instead we
must take the first/last statement for the given index. We can use
psearch/nsearch for it - it will position to the first/last element in
the tree equal to the search key; since the search key is equal to any
statement of the given index in case the given key is empty, it will do
the job.

While we are at it, also remove handling 'key == NULL' case from the
write_set_key_cmp() as it is not used anywhere.