Revert "vinyl: implement deferred restore of runs and mem"
This reverts commit 1360f193435fcd7c3b6b84ad9264176fc7718527.

The idea behind this patch makes sense for memory iterator, but the
implementation is incorrect: it's not enough to patch ->restore, because
memory iterator can start right from ->next_key, bypassing ->restore.
Since other hunks of this patch are not needed in the scope of the run
slice paradigm, and what is done by this patch needs to be rewritten
from scratch anyway, revert the whole patch.