Rework vinyl/recover test
The vinyl/recover test was written long time ago. Back then recovery was
a convoluted procedure based on scanning the data directory so that it
was crucial to generate a number of stale files to check its validity.
Nowadays, there's no point in it thanks to the metadata log. Moreover,
when the single memory level is introduced, ERRINJ_VY_RANGE_SPLIT won't
make sense any more as range splitting will not involve a worker thread.
So we can remove errinj from this test.

Another problem with this test is that it doesn't take into account data
compression (all tuples generated by it are compressed perfectly).
Handle this by generating random padding strings.

Also, when we snapshot vinyl stats before restart, there still may be
compaction in progress that will modify the stats in the last moment,
resulting in a sporadic failure. Address that by checking stats
separately, for another space with compaction disabled.