vinyl: don't create empty slices on dump
Adding an empty slice to a range is pointless, besides it triggers
compaction for no reason, which is especially harmful in case of
time-series-like workload. On dump we can omit creating slices for
ranges that are not intersected by the new run. Note how it affects the
coalesce test: now we have to insert a statement into each range to
trigger compaction, not just into the first one.